meta-agent-eval-system/pyproject.toml
[project]
name = "meta-agent-eval-system"
version = "0.1.0"
description = "Multi-turn eval pipeline and incident-to-eval taxonomy"
requires-python = "==3.11.*"
dependencies = [
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "openai>=2.7.1",
    "streamlit>=1.50.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyarrow>=21.0.0",  # Pin to version with pre-built wheels
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "eval"]
